newsportal/web/src/main/java/by/pvt/herzhot/controller/Controller.java
package by.pvt.herzhot.controller;

import by.pvt.herzhot.commands.ICommand;
import by.pvt.herzhot.commands.factory.CommandFactory;
import by.pvt.herzhot.constants.ConfigConstants;
import by.pvt.herzhot.managers.impl.ConfigManagerImpl;

import javax.servlet.RequestDispatcher;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

/**
 * @author Herzhot
 * @version 1.0
 *          09.05.2016
 */
public class Controller extends HttpServlet {

    private static final long serialVersionUID = 1L;

    protected void processRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        String page;
        ICommand command = CommandFactory.INSTANCE.defineCommand(request);
        page = command.execute(request);
        if(page != null){
            RequestDispatcher dispatcher = request.getServletContext().getRequestDispatcher(page);
            dispatcher.forward(request, response);
        }
        else{
            page = ConfigManagerImpl.INSTANCE.getProperty(ConfigConstants.INDEX_PAGE_PATH);
            RequestDispatcher dispatcher = request.getServletContext().getRequestDispatcher(page);
            dispatcher.forward(request, response);
        }
    }

    @Override
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        processRequest(request, response);
    }

    @Override
    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        processRequest(request, response);
    }

}
